+++ iGlance/iGlance/iGlance/MenuBarItems/DiskUsageMenuBarItem.swift
@@ -76,7 +76,7 @@ class DiskUsageMenuBarItem: MenuBarItem {
         let menuBarImage = NSImage(
             size: NSSize(
                 width: iconWidth + textWidth,
-                height: CGFloat(self.menuBarHeight)
+                height: self.menuBarHeight
             )
         )
